+++ src/cli/start.ts
@@ -10,8 +10,7 @@ export default (unloggedContext: UnloggedContext) => {
 			const context = unloggedContext.withArgv(argv);
 
 			try {
-				const instance = await createInstance(context, false);
-				await instance.deploy();
+				const instance = await createInstance(context, { attach: false });
 				await instance.detach();
 				instance.announce();
 			} catch (error) {