+++ src/context/index.ts
@@ -5,12 +5,15 @@ import { sync as findUp } from "find-up";
 
 import { CommonArgv } from "../cli";
 import { createLogger, LogLevel } from "./logger";
-import { normalizeRc, NormalizedRc } from "./rc";
+import { normalizeRc, NormalizedRc, Deployment, NanoDeployment } from "./rc";
+import { get as remoteNano } from "@crkn-rcdr/nano";
+import { getInstance } from "../instance";
 
 export { logLevels, LogLevel } from "./logger";
 
 export {
 	Deployment,
+	NanoDeployment,
 	InstanceConfig,
 	normalizeInstanceConfig,
 	NormalizedInstanceConfig,
@@ -24,6 +27,10 @@ export type UnloggedContext = NormalizedRc & {
 
 export type Context = Omit<UnloggedContext, "withArgv"> & {
 	readonly log: (level: LogLevel, message: string) => void;
+	readonly getDeployment: (
+		key: string,
+		suffix?: string
+	) => Promise<NanoDeployment>;
 	readonly withDatabase: (db: string) => DatabaseContext;
 };
 
@@ -42,16 +49,44 @@ export const createContext = (directory: string): UnloggedContext => {
 	return {
 		directory,
 		...rc,
-		withArgv: function (argv: CommonArgv): Context {
+		withArgv: function (argv: CommonArgv) {
 			// https://no-color.org
 			if (process.env.hasOwnProperty("NO_COLOR")) argv.color = false;
 
 			const logger = createLogger(argv);
+			if (!confPath)
+				logger.log(
+					"warn",
+					"No kivikrc file detected. Proceeding with defaults."
+				);
 			logger.log("info", "Logger initialized.");
 
 			return {
 				...this,
 				log: (level: LogLevel, message: string) => logger.log(level, message),
+				getDeployment: async function (key: string, suffix?: string) {
+					if (key in this.deployments) {
+						const deployment = this.deployments[key] as Deployment;
+						return {
+							nano: remoteNano(deployment.url, deployment.auth),
+							suffix: suffix || deployment.suffix,
+							fixtures: !!deployment.fixtures,
+							dbs: deployment.dbs || null,
+						};
+					} else if (key === "local") {
+						const instance = await getInstance(this);
+						return {
+							nano: instance.nano,
+							suffix,
+							fixtures: this.local.fixtures,
+							dbs: null,
+						};
+					} else {
+						throw new Error(
+							`Your kivikrc file does not have a deployment with key '${key}'`
+						);
+					}
+				},
 				withDatabase: function (db: string): DatabaseContext {
 					return {
 						...this,