+++ src/instance/index.ts
@@ -1,36 +1,50 @@
+import { ServerScope } from "nano";
+
+import { Context, defaultContext } from "../context";
 import {
-	Context,
-	defaultContext,
-	InstanceConfig,
-	NormalizedInstanceConfig,
-	normalizeInstanceConfig,
-} from "../context";
-import { createKivik, DatabaseHandlerMap, Kivik } from "../kivik";
+	createKivik,
+	DatabaseHandler,
+	DatabaseHandlerMap,
+	Kivik,
+} from "../kivik";
 import { getContainer, createContainer, Container } from "./container";
+
+export interface InstanceOptions {
+	/**
+	 * Whether to attach to a running instance when trying to create one.
+	 * Ignored when trying to get a running instance directly.
+	 * Default: true
+	 */
+	attach?: boolean;
+	/**
+	 * Whether to deploy fixtures to the instance when deploying design
+	 * documents. Default: Whatever is set in `local.fixtures` in your RC file.
+	 * If that's unset, the default is `true`.
+	 */
+	fixtures?: boolean;
+}
 
 /**
  * Gets a running Kivik instance. Throws an error if the instance cannot be
  * found.
  * @param directory The root directory for the files Kivik will manage.
- * @param config Instance configuration.
  */
 export async function getInstance(
 	directory: string,
-	config?: InstanceConfig
+	options?: InstanceOptions
 ): Promise<Instance>;
 /**
  * Gets a running Kivik instance. Throws an error if the instance cannot be
  * found.
  * @param context The Kivik context object to apply to the instance.
- * @param config Instance configuration.
  */
 export async function getInstance(
 	context: Context,
-	config?: InstanceConfig
+	options?: InstanceOptions
 ): Promise<Instance>;
 export async function getInstance(
 	input: string | Context,
-	config: InstanceConfig = {}
+	options: InstanceOptions = {}
 ): Promise<Instance> {
 	const context = typeof input === "string" ? defaultContext(input) : input;
 	const container = await getContainer(context);
@@ -38,51 +52,43 @@ export async function getInstance(
 		throw new Error(
 			"Cannot find a running instance. Check that .kivik.tmp exists and contains the name of a running Docker container."
 		);
-	const nConfig = normalizeInstanceConfig(
-		Object.assign({}, context.local, config)
-	);
-	return await instanceHelper(container, context, nConfig);
+
+	return await instanceHelper(container, context, options);
 }
 
 /**
  * Creates a Kivik instance.
  * @param directory The root directory for the files Kivik will manage.
- * @param attach Attach to a running instance, if it exists.
- * @param config Instance configuration.
  * @returns The Kivik instance. File scan and load is complete, and the Docker
  * container running the instance's CouchDB endpoint is ready to go.
  */
 export async function createInstance(
 	directory: string,
-	attach?: boolean,
-	config?: InstanceConfig
+	options?: InstanceOptions
 ): Promise<Instance>;
 /**
  * Creates a Kivik instance.
  * @param context The Kivik context object to apply to the instance.
- * @param attach Attach to a running instance, if it exists.
- * @param config Instance configuration.
  * @returns The Kivik instance. File scan and load is complete, and the Docker
  * container running the instance's CouchDB endpoint is ready to go.
  */
 export async function createInstance(
 	context: Context,
-	attach?: boolean,
-	config?: InstanceConfig
+	options?: InstanceOptions
 ): Promise<Instance>;
 export async function createInstance(
 	input: string | Context,
-	attach = true,
-	config: InstanceConfig = {}
+	options: InstanceOptions = {}
 ) {
 	const context = typeof input === "string" ? defaultContext(input) : input;
 
 	let instance: Instance | null = null;
 	try {
-		instance = await getInstance(context, config);
+		instance = await getInstance(context, options);
 	} catch (_) {}
 
 	if (instance) {
+		const attach = "attach" in options ? !!options.attach : true;
 		if (attach) {
 			context.log("success", "Attaching to a running instance.");
 			return instance;
@@ -96,29 +102,27 @@ export async function createInstance(
 		);
 	}
 
-	const nConfig = normalizeInstanceConfig(
-		Object.assign({}, context.local, config)
-	);
-	const container = await createContainer(context, nConfig);
-	return await instanceHelper(container, context, nConfig);
+	const container = await createContainer(context);
+	return await instanceHelper(container, context, options);
 }
 
 async function instanceHelper(
 	container: Container,
 	context: Context,
-	config: NormalizedInstanceConfig
+	options: InstanceOptions
 ): Promise<Instance> {
-	const nConfig = normalizeInstanceConfig(
-		Object.assign({}, context.local, config)
-	);
+	const fixtures =
+		"fixtures" in options
+			? !!options.fixtures
+			: "fixtures" in context.local
+			? !!context.local.fixtures
+			: true;
 
-	const kivik = await createKivik(
-		context,
-		nConfig.fixtures ? "instance" : "deploy"
-	);
+	const kivik = await createKivik(context, fixtures ? "instance" : "deploy");
 
 	return {
 		kivik,
+		nano: container.nano,
 		announce: () => {
 			container.announce();
 		},
@@ -127,7 +131,22 @@ async function instanceHelper(
 			kivik.deployOnChanges(container.nano);
 		},
 		deploy: async (suffix?: string) => {
-			return await kivik.deploy(container.nano, suffix);
+			return await kivik.deploy({
+				nano: container.nano,
+				suffix,
+				fixtures,
+				dbs: null,
+			});
+		},
+		deployDb: async (db: string, suffix?: string) => {
+			const handlers = await kivik.deploy({
+				nano: container.nano,
+				suffix,
+				fixtures,
+				dbs: [db],
+			});
+			if (!handlers.has(db)) throw new Error(`Database '${db}' not found.`);
+			return handlers.get(db) as DatabaseHandler;
 		},
 		detach: async () => {
 			await kivik.close();
@@ -142,9 +161,11 @@ async function instanceHelper(
 
 export interface Instance {
 	readonly kivik: Kivik;
+	readonly nano: ServerScope;
 	readonly announce: () => void;
 	readonly attach: () => Promise<void>;
 	readonly deploy: (suffix?: string) => Promise<DatabaseHandlerMap>;
+	readonly deployDb: (db: string, suffix?: string) => Promise<DatabaseHandler>;
 	readonly detach: () => Promise<void>;
 	readonly stop: () => Promise<void>;
 }